Add a custom field to a portfolio

Custom fields are associated with portfolios by way of custom field settings. This method creates a setting for the portfolio.

Usage

asn_add_custom_field_setting_for_portfolio(portfolio_gid)

Arguments

portfolio_gid

(character scalar) Globally unique identifier for the portfolio.

Value

Custom Fields Settings objects represent the many-to-many join of the Custom Field and Project as well as stores information that is relevant to that particular pairing.

Property	Class	Description

gid	character scalar	Globally unique identifier of the resource, as a string.

resource_type	character scalar	The base type of this resource.

custom_field	list	The custom field that is applied to the parent.

is_important	logical scalar	is_important is used in the Asana web application to determine if this custom field is displayed in the list/grid view of a project or portfolio.

parent	list	The parent to which the custom field is applied. This can be a project or portfolio and indicates that the tasks or projects that the parent contains may be given custom field values for this custom field.

project	list	Deprecated: new integrations should prefer the parent field. The id of the project that this custom field settings refers to.