Add followers to a project

Adds the specified list of users as followers to the project. Followers are a subset of members who have opted in to receive "tasks added" notifications for a project. Therefore, if the users are not already members of the project, they will also become members as a result of this operation. Returns the updated project record.

Usage

asn_add_followers_for_project(project_gid, opt_fields)

Arguments

project_gid

(character scalar) Globally unique identifier for the project.

opt_fields

(optional) (list) Defines fields to return. Some requests return compact representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.

Value

A project represents a prioritized list of tasks in Asana or a board with columns of tasks represented as cards. It exists in a single workspace or organization and is accessible to a subset of users in that workspace or organization, depending on its permissions.

Property	Class	Description

gid	character scalar	Globally unique identifier of the resource, as a string.

resource_type	character scalar	The base type of this resource.

name	character scalar	Name of the project. This is generally a short sentence fragment that fits on a line in the UI for maximum readability. However, it can be longer.

archived	logical scalar	True if the project is archived, false if not. Archived projects do not show in the UI by default and may be treated differently for queries.

color	character scalar	Color of the project.

created_at	datetime scalar	The time at which this resource was created.

current_status	list	Deprecated: new integrations should prefer the current_status_update resource.

current_status_update	list	The latest status_update posted to this project.

custom_field_settings	list	Array of Custom Field Settings (in compact form).

default_view	character scalar	The default view (list, board, calendar, or timeline) of a project.

due_date	datetime scalar	Deprecated: new integrations should prefer the due_on field.

due_on	datetime scalar	The day on which this project is due. This takes a date with format YYYY-MM-DD.

html_notes	character scalar	Opt In. The notes of the project with formatting as HTML.

members	list	Array of users who are members of this project.

modified_at	datetime scalar	The time at which this project was last modified. Note: This does not currently reflect any changes in associations such as tasks or comments that may have been added or removed from the project.

notes	character scalar	Free-form textual information associated with the project (ie., its description).

public	logical scalar	True if the project is public to its team.

start_on	date scalar	The day on which work for this project begins, or null if the project has no start date. This takes a date with YYYY-MM-DD format. Note: due_on or due_at must be present in the request when setting or unsetting the start_on parameter. Additionally, start_on and due_on cannot be the same date.

workspace	list	Create-only. The workspace or organization this project is associated with. Once created, projects cannot be moved to a different workspace. This attribute can only be specified at creation time.

completed	logical scalar	True if the project is currently marked complete, false if not.

completed_at	datetime scalar	The time at which this project was completed, or null if the project is not completed.

completed_by	list	The user that marked this project complete, or null if the project is not completed.

created_from_template	list	Opt In. The project template from which this project was created. If the project was not created from a template, this field will be null.

custom_fields	list	Array of Custom Fields.

followers	list	Array of users following this project. Followers are a subset of members who have opted in to receive "tasks added" notifications for a project.

icon	character scalar	The icon for a project.

owner	list	The current owner of the project, may be null.

permalink_url	character scalar	A url that points directly to the object within Asana.

project_brief	list	Opt In. The project brief associated with this project.

team	list	The team that this project is shared with.