Create an enum option

Creates an enum option and adds it to this custom field’s list of enum options. A custom field can have at most 500 enum options (including disabled options). By default new enum options are inserted at the end of a custom field’s list. Locked custom fields can only have enum options added by the user who locked the field. Returns the full record of the newly created enum option.

Usage

asn_create_enum_option_for_custom_field(
  custom_field_gid,
  limit,
  offset,
  opt_fields
)

Arguments

custom_field_gid

(character scalar) Globally unique identifier for the custom field.

limit

(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.

offset

(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'

opt_fields

(optional) (list) Defines fields to return. Some requests return compact representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.

Value

Enum options are the possible values which an enum custom field can adopt. An enum custom field must contain at least 1 enum option but no more than 500. You can add enum options to a custom field by using the POST /custom_fields/custom_field_gid/enum_options endpoint. It is not possible to remove or delete an enum option. Instead, enum options can be disabled by updating the enabled field to false with the PUT /enum_options/enum_option_gid endpoint. Other attributes can be updated similarly. On creation of an enum option, enabled is always set to true, meaning the enum option is a selectable value for the custom field. Setting enabled=false is equivalent to “trashing” the enum option in the Asana web app within the “Edit Fields” dialog. The enum option will no longer be selectable but, if the enum option value was previously set within a task, the task will retain the value. Enum options are an ordered list and by default new enum options are inserted at the end. Ordering in relation to existing enum options can be specified on creation by using insert_before or insert_after to reference an existing enum option. Only one of insert_before and insert_after can be provided when creating a new enum option. An enum options list can be reordered with the POST /custom_fields/custom_field_gid/enum_options/insert endpoint.

Property	Class	Description

gid	character scalar	Globally unique identifier of the resource, as a string.

resource_type	character scalar	The base type of this resource.

color	character scalar	The color of the enum option. Defaults to ‘none’.

enabled	logical scalar	Whether or not the enum option is a selectable value for the custom field.

name	character scalar	The name of the enum option.