Get an attachment

Get the full record for a single attachment.

Usage

asn_get_attachment(attachment_gid, opt_fields)

Arguments

attachment_gid

(character scalar) Globally unique identifier for the attachment.

opt_fields

(optional) (list) Defines fields to return. Some requests return compact representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.

Value

An attachment object represents any file attached to a task in Asana, whether it’s an uploaded file or one associated via a third-party service such as Dropbox or Google Drive.

Property	Class	Description

gid	character scalar	Globally unique identifier of the resource, as a string.

resource_type	character scalar	The base type of this resource.

name	character scalar	The name of the file.

resource_subtype	character scalar	The service hosting the attachment. Valid values are asana, dropbox, gdrive, onedrive, box, vimeo, and external.

connected_to_app	logical scalar	Whether the attachment is connected to the app making the request for the purposes of showing an app components widget. Only present when the resource_subtype is external or gdrive.

created_at	datetime scalar	The time at which this resource was created.

download_url	character scalar	The URL containing the content of the attachment. Note: May be null if the attachment is hosted by Box and will be null if the attachment is a Video Message hosted by Vimeo. If present, this URL may only be valid for two minutes from the time of retrieval. You should avoid persisting this URL somewhere and just refresh it on demand to ensure you do not keep stale URLs.

host	character scalar	The service hosting the attachment. Valid values are asana, dropbox, gdrive, box, and vimeo.

parent	list	The task this attachment is attached to.

permanent_url	character scalar	NA

size	integer scalar	The size of the attachment in bytes. Only present when the resource_subtype is asana.

view_url	character scalar	The URL where the attachment can be viewed, which may be friendlier to users in a browser than just directing them to a raw file. May be null if no view URL exists for the service.