Get a workspace's custom fields

Returns a list of the compact representation of all of the custom fields in a workspace.

Usage

asn_get_custom_fields_for_workspace(workspace_gid, limit, offset, opt_fields)

Arguments

workspace_gid

(character scalar) Globally unique identifier for the workspace or organization.

limit

(optional) (integer scalar) Results per page. The number of objects to return per page. The value must be between 1 and 100.

offset

(optional) (character scalar) Offset token. An offset to the next page returned by the API. A pagination request will return an offset token, which can be used as an input parameter to the next request. If an offset is not passed in, the API will return the first page of results. 'Note: You can only pass in an offset that was returned to you via a previously paginated request.'

opt_fields

(optional) (list) Defines fields to return. Some requests return compact representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.

Value

Custom Fields store the metadata that is used in order to add user-specified information to tasks in Asana. Be sure to reference the custom fields developer documentation for more information about how custom fields relate to various resources in Asana. Users in Asana can lock custom fields, which will make them read-only when accessed by other users. Attempting to edit a locked custom field will return HTTP error code 403 Forbidden.

Property	Class	Description

gid	character scalar	Globally unique identifier of the resource, as a string.

resource_type	character scalar	The base type of this resource.

date_value	list	Conditional. Only relevant for custom fields of type date. This object reflects the chosen date (and optionally, time) value of a date custom field. If no date is selected, the value of date_value will be null.

display_value	character scalar	A string representation for the value of the custom field. Integrations that don't require the underlying type should use this field to read values. Using this field will future-proof an app against new custom field types.

enabled	logical scalar	Conditional. Determines if the custom field is enabled or not.

enum_options	list	Conditional. Only relevant for custom fields of type enum. This array specifies the possible values which an enum custom field can adopt. To modify the enum options, refer to working with enum options.

enum_value	list	Conditional. Only relevant for custom fields of type enum. This object is the chosen value of an enum custom field.

multi_enum_values	list	Conditional. Only relevant for custom fields of type multi_enum. This object is the chosen values of a multi_enum custom field.

name	character scalar	The name of the custom field.

number_value	number	Conditional. This number is the value of a number custom field.

resource_subtype	character scalar	The type of the custom field. Must be one of the given values.

text_value	character scalar	Conditional. This string is the value of a text custom field.

type	character scalar	Deprecated: new integrations should prefer the resource_subtype field. The type of the custom field. Must be one of the given values.

asana_created_field	character scalar	Conditional. A unique identifier to associate this field with the template source of truth.

currency_code	character scalar	ISO 4217 currency code to format this custom field. This will be null if the format is not currency.

custom_label	character scalar	This is the string that appears next to the custom field value. This will be null if the format is not custom.

custom_label_position	character scalar	Only relevant for custom fields with custom format. This depicts where to place the custom label. This will be null if the format is not custom.

description	character scalar	Opt In. The description of the custom field.

format	character scalar	The format of this custom field.

has_notifications_enabled	logical scalar	Conditional. This flag describes whether a follower of a task with this field should receive inbox notifications from changes to this field.

is_global_to_workspace	logical scalar	This flag describes whether this custom field is available to every container in the workspace. Before project-specific custom fields, this field was always true.

precision	integer scalar	Only relevant for custom fields of type ‘Number’. This field dictates the number of places after the decimal to round to, i.e. 0 is integer values, 1 rounds to the nearest tenth, and so on. Must be between 0 and 6, inclusive. For percentage format, this may be unintuitive, as a value of 0.25 has a precision of 0, while a value of 0.251 has a precision of 1. This is due to 0.25 being displayed as 25%. The identifier format will always have a precision of 0.

created_by	list	A user object represents an account in Asana that can be given access to various workspaces, projects, and tasks.

people_value	list	Conditional. Only relevant for custom fields of type people. This array of compact user objects reflects the values of a people custom field.