Get events on a resource

Returns the full record for all events that have occurred since the sync token was created. A GET request to the endpoint /[path_to_resource]/events can be made in lieu of including the resource ID in the data for the request. Asana limits a single sync token to 100 events. If more than 100 events exist for a given resource, has_more: true will be returned in the response, indicating that there are more events to pull. Note: The resource returned will be the resource that triggered the event. This may be different from the one that the events were requested for. For example, a subscription to a project will contain events for tasks contained within the project.

Usage

asn_get_events(resource, opt_fields, sync)

Arguments

resource

(character scalar) A resource ID to subscribe to. The resource can be a task or project.

opt_fields

(optional) (list) Defines fields to return. Some requests return compact representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.

sync

(optional) (character scalar) A sync token received from the last request, or none on first sync. Events will be returned from the point in time that the sync token was generated. Note: On your first request, omit the sync token. The response will be the same as for an expired sync token, and will include a new valid sync token.If the sync token is too old (which may happen from time to time) the API will return a 412 Precondition Failed error, and include a fresh sync token in the response.

Value

An event is an object representing a change to a resource that was observed by an event subscription or delivered asynchronously to the target location of an active webhook. The event may be triggered by a different user than the subscriber. For example, if user A subscribes to a task and user B modified it, the event’s user will be user B. Note: Some events are generated by the system, and will have null as the user. API consumers should make sure to handle this case. The resource that triggered the event may be different from the one that the events were requested for or the webhook is subscribed to. For example, a subscription to a project will contain events for tasks contained within the project. Note: pay close attention to the relationship between the fields Event.action and Event.change.action. Event.action represents the action taken on the resource itself, and Event.change.action represents how the information within the resource's fields have been modified. For instance, consider these scenarios: * When at task is added to a project, Event.action will be added, Event.parent will be an object with the id and type of the project, and there will be no change field. * When an assignee is set on the task, Event.parent will be null, Event.action will be changed, Event.change.action will be changed, and new_value will be an object with the user's id and type. * When a collaborator is added to the task, Event.parent will be null, Event.action will be changed, Event.change.action will be added, and added_value will be an object with the user's id and type.

Property	Class	Description

action	character scalar	The type of action taken on the resource that triggered the event. This can be one of changed, added, removed, deleted, or undeleted depending on the nature of the event.

change	list	Information about the type of change that has occurred. This field is only present when the value of the property action, describing the action taken on the resource, is changed.

created_at	datetime scalar	The timestamp when the event occurred.

parent	list	For added/removed events, the parent object that resource was added to or removed from. The parent will be null for other event types.

resource	list	The resource which has triggered the event by being modified in some way.

type	character scalar	Deprecated: Refer to the resource_type of the resource. The type of the resource that generated the event.

user	list	The user who triggered the event.