Get a goal relationship

Returns the complete updated goal relationship record for a single goal relationship.

Usage

asn_get_goal_relationship(goal_relationship_gid, opt_fields)

Arguments

goal_relationship_gid

(character scalar) Globally unique identifier for the goal relationship.

opt_fields

(optional) (list) Defines fields to return. Some requests return compact representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.

Value

A goal relationship is an object representing the relationship between a goal and another goal, a project, or a portfolio.

Property	Class	Description

gid	character scalar	Globally unique identifier of the resource, as a string.

resource_type	character scalar	The base type of this resource.

contribution_weight	number	The weight that the supporting resource's progress contributes to the supported goal's progress. This can only be 0 or 1.

resource_subtype	character scalar	The subtype of this resource. Different subtypes retain many of the same fields and behavior, but may render differently in Asana or represent resources with different semantic meaning.

supporting_resource	list	The supporting resource that supports the goal. This can be either a project, portfolio, or goal.

supported_goal	list	The goal that the supporting resource supports.