Get a portfolio

Returns the complete portfolio record for a single portfolio.

Usage

asn_get_portfolio(portfolio_gid, opt_fields)

Arguments

portfolio_gid

(character scalar) Globally unique identifier for the portfolio.

opt_fields

(optional) (list) Defines fields to return. Some requests return compact representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.

Value

A portfolio gives a high-level overview of the status of multiple initiatives in Asana. Portfolios provide a dashboard overview of the state of multiple projects, including a progress report and the most recent project status update. Portfolios have some restrictions on size. Each portfolio has a max of 500 items and, like projects, a max of 20 custom fields.

Property	Class	Description

gid	character scalar	Globally unique identifier of the resource, as a string.

resource_type	character scalar	The base type of this resource.

name	character scalar	The name of the portfolio.

color	character scalar	Color of the portfolio.

created_at	datetime scalar	The time at which this resource was created.

created_by	list	A user object represents an account in Asana that can be given access to various workspaces, projects, and tasks.

current_status_update	list	The latest status_update posted to this portfolio.

custom_field_settings	list	Array of custom field settings applied to the portfolio.

custom_fields	list	Array of Custom Fields.

due_on	datetime scalar	The localized day on which this portfolio is due. This takes a date with format YYYY-MM-DD.

members	list	A user object represents an account in Asana that can be given access to various workspaces, projects, and tasks.

owner	list	A user object represents an account in Asana that can be given access to various workspaces, projects, and tasks.

permalink_url	character scalar	A url that points directly to the object within Asana.

public	logical scalar	True if the portfolio is public to its workspace members.

start_on	date scalar	The day on which work for this portfolio begins, or null if the portfolio has no start date. This takes a date with YYYY-MM-DD format. Note: due_on must be present in the request when setting or unsetting the start_on parameter. Additionally, start_on and due_on cannot be the same date.

workspace	list	Create-only. The workspace or organization that the portfolio belongs to.