Search tasks in a workspace

To mirror the functionality of the Asana web app's advanced search feature, the Asana API has a task search endpoint that allows you to build complex filters to find and retrieve the exact data you need. #### Premium access Like the Asana web product's advance search feature, this search endpoint will only be available to premium Asana users. A user is premium if any of the following is true: - The workspace in which the search is being performed is a premium workspace - The user is a member of a premium team inside the workspace Even if a user is only a member of a premium team inside a non-premium workspace, search will allow them to find data anywhere in the workspace, not just inside the premium team. Making a search request using credentials of a non-premium user will result in a 402 Payment Required error. #### Pagination Search results are not stable; repeating the same query multiple times may return the data in a different order, even if the data do not change. Because of this, the traditional pagination available elsewhere in the Asana API is not available here. However, you can paginate manually by sorting the search results by their creation time and then modifying each subsequent query to exclude data you have already seen. Page sizes are limited to a maximum of 100 items, and can be specified by the limit query parameter. #### Eventual consistency Changes in Asana (regardless of whether they’re made though the web product or the API) are forwarded to our search infrastructure to be indexed. This process can take between 10 and 60 seconds to complete under normal operation, and longer during some production incidents. Making a change to a task that would alter its presence in a particular search query will not be reflected immediately. This is also true of the advanced search feature in the web product. #### Rate limits You may receive a 429 Too Many Requests response if you hit any of our rate limits. #### Custom field parameters | Parameter name | Custom field type | Accepted type | |---|---|---| | custom_fields.gid.is_set | All | Boolean | | custom_fields.gid.value | Text | String | | custom_fields.gid.value | Number | Number | | custom_fields.gid.value | Enum | Enum option ID | | custom_fields.gid.starts_with | Text only | String | | custom_fields.gid.ends_with | Text only | String | | custom_fields.gid.contains | Text only | String | | custom_fields.gid.less_than | Number only | Number | | custom_fields.gid.greater_than | Number only | Number | For example, if the gid of the custom field is 12345, these query parameter to find tasks where it is set would be custom_fields.12345.is_set=true. To match an exact value for an enum custom field, use the gid of the desired enum option and not the name of the enum option: custom_fields.12345.value=67890. Not Supported: searching for multiple exact matches of a custom field, searching for multi-enum custom field Note: If you specify projects.any and sections.any, you will receive tasks for the project and tasks for the section. If you're looking for only tasks in a section, omit the projects.any from the request.

Usage

asn_search_tasks_for_workspace(
  workspace_gid,
  assigned_by.any,
  assigned_by.not,
  assignee.any,
  assignee.not,
  commented_on_by.not,
  completed,
  completed_at.after,
  completed_at.before,
  completed_on,
  completed_on.after,
  completed_on.before,
  created_at.after,
  created_at.before,
  created_by.any,
  created_by.not,
  created_on,
  created_on.after,
  created_on.before,
  due_at.after,
  due_at.before,
  due_on,
  due_on.after,
  due_on.before,
  followers.not,
  has_attachment,
  is_blocked,
  is_blocking,
  is_subtask,
  liked_by.not,
  modified_at.after,
  modified_at.before,
  modified_on,
  modified_on.after,
  modified_on.before,
  opt_fields,
  portfolios.any,
  projects.all,
  projects.any,
  projects.not,
  resource_subtype = "milestone",
  sections.all,
  sections.any,
  sections.not,
  sort_ascending = "FALSE",
  sort_by = "modified_at",
  start_on,
  start_on.after,
  start_on.before,
  tags.all,
  tags.any,
  tags.not,
  teams.any,
  text
)

Arguments

workspace_gid

(character scalar) Globally unique identifier for the workspace or organization.

assigned_by.any

(optional) (character scalar) Comma-separated list of user identifiers

assigned_by.not

(optional) (character scalar) Comma-separated list of user identifiers

assignee.any

(optional) (character scalar) Comma-separated list of user identifiers

assignee.not

(optional) (character scalar) Comma-separated list of user identifiers

commented_on_by.not

(optional) (character scalar) Comma-separated list of user identifiers

completed

(optional) (logical scalar) Filter to completed tasks

completed_at.after

(optional) (datetime scalar) ISO 8601 datetime string

completed_at.before

(optional) (datetime scalar) ISO 8601 datetime string

completed_on

(optional) (date scalar) ISO 8601 date string or null

completed_on.after

(optional) (date scalar) ISO 8601 date string

completed_on.before

(optional) (date scalar) ISO 8601 date string

created_at.after

(optional) (datetime scalar) ISO 8601 datetime string

created_at.before

(optional) (datetime scalar) ISO 8601 datetime string

created_by.any

(optional) (character scalar) Comma-separated list of user identifiers

created_by.not

(optional) (character scalar) Comma-separated list of user identifiers

created_on

(optional) (date scalar) ISO 8601 date string or null

created_on.after

(optional) (date scalar) ISO 8601 date string

created_on.before

(optional) (date scalar) ISO 8601 date string

due_at.after

(optional) (datetime scalar) ISO 8601 datetime string

due_at.before

(optional) (datetime scalar) ISO 8601 datetime string

due_on

(optional) (date scalar) ISO 8601 date string or null

due_on.after

(optional) (date scalar) ISO 8601 date string

due_on.before

(optional) (date scalar) ISO 8601 date string

followers.not

(optional) (character scalar) Comma-separated list of user identifiers

has_attachment

(optional) (logical scalar) Filter to tasks with attachments

is_blocked

(optional) (logical scalar) Filter to tasks with incomplete dependencies

is_blocking

(optional) (logical scalar) Filter to incomplete tasks with dependents

is_subtask

(optional) (logical scalar) Filter to subtasks

liked_by.not

(optional) (character scalar) Comma-separated list of user identifiers

modified_at.after

(optional) (datetime scalar) ISO 8601 datetime string

modified_at.before

(optional) (datetime scalar) ISO 8601 datetime string

modified_on

(optional) (date scalar) ISO 8601 date string or null

modified_on.after

(optional) (date scalar) ISO 8601 date string

modified_on.before

(optional) (date scalar) ISO 8601 date string

opt_fields

(optional) (character vector) Defines fields to return. Some requests return compact representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.

portfolios.any

(optional) (character scalar) Comma-separated list of portfolio IDs

projects.all

(optional) (character scalar) Comma-separated list of project IDs

projects.any

(optional) (character scalar) Comma-separated list of project IDs

projects.not

(optional) (character scalar) Comma-separated list of project IDs

resource_subtype

(optional) (character scalar) Filters results by the task's resource_subtype Possible values: default_task, milestone. Default: milestone

sections.all

(optional) (character scalar) Comma-separated list of section or column IDs

sections.any

(optional) (character scalar) Comma-separated list of section or column IDs

sections.not

(optional) (character scalar) Comma-separated list of section or column IDs

sort_ascending

(optional) (logical scalar) Default false Default: FALSE

sort_by

(optional) (character scalar) One of due_date, created_at, completed_at, likes, or modified_at, defaults to modified_at Possible values: due_date, created_at, completed_at, likes, modified_at. Default: modified_at

start_on

(optional) (date scalar) ISO 8601 date string or null

start_on.after

(optional) (date scalar) ISO 8601 date string

start_on.before

(optional) (date scalar) ISO 8601 date string

tags.all

(optional) (character scalar) Comma-separated list of tag IDs

tags.any

(optional) (character scalar) Comma-separated list of tag IDs

tags.not

(optional) (character scalar) Comma-separated list of tag IDs

teams.any

(optional) (character scalar) Comma-separated list of team IDs

text

(optional) (character scalar) Performs full-text search on both task name and description

Value

A generic Asana Resource, containing a globally unique identifier.