Update a task

A specific, existing task can be updated by making a PUT request on the URL for that task. Only the fields provided in the data block will be updated; any unspecified fields will remain unchanged. When using this method, it is best to specify only those fields you wish to change, or else you may overwrite changes made by another user since you last retrieved the task. Returns the complete updated task record.

Usage

asn_update_task(task_gid, opt_fields)

Arguments

task_gid

(character scalar) The task to operate on.

opt_fields

(optional) (list) Defines fields to return. Some requests return compact representations of objects in order to conserve resources and complete the request more efficiently. Other times requests return more information than you may need. This option allows you to list the exact set of fields that the API should be sure to return for the objects. The field names should be provided as paths, described below. The id of included objects will always be returned, regardless of the field options.

Value

The task is the basic object around which many operations in Asana are centered.

Property	Class	Description

gid	character scalar	Globally unique identifier of the resource, as a string.

resource_type	character scalar	The base type of this resource.

name	character scalar	The name of the task.

resource_subtype	character scalar	The subtype of this resource. Different subtypes retain many of the same fields and behavior, but may render differently in Asana or represent resources with different semantic meaning. The resource_subtype milestone represent a single moment in time. This means tasks with this subtype cannot have a start_date.

actual_time_minutes	number	This value represents the sum of all the Time Tracking entries in the Actual Time field on a given Task. It is represented as a nullable long value.

approval_status	character scalar	Conditional Reflects the approval status of this task. This field is kept in sync with completed, meaning pending translates to false while approved, rejected, and changes_requested translate to true. If you set completed to true, this field will be set to approved.

assignee_status	character scalar	Deprecated Scheduling status of this task for the user it is assigned to. This field can only be set if the assignee is non-null. Setting this field to "inbox" or "upcoming" inserts it at the top of the section, while the other options will insert at the bottom.

completed	logical scalar	True if the task is currently marked complete, false if not.

completed_at	datetime scalar	The time at which this task was completed, or null if the task is incomplete.

completed_by	list	A user object represents an account in Asana that can be given access to various workspaces, projects, and tasks.

created_at	datetime scalar	The time at which this resource was created.

dependencies	list	Opt In. Array of resources referencing tasks that this task depends on. The objects contain only the gid of the dependency.

dependents	list	Opt In. Array of resources referencing tasks that depend on this task. The objects contain only the ID of the dependent.

due_at	date scalar	The UTC date and time on which this task is due, or null if the task has no due time. This takes an ISO 8601 date string in UTC and should not be used together with due_on.

due_on	date scalar	The localized date on which this task is due, or null if the task has no due date. This takes a date with YYYY-MM-DD format and should not be used together with due_at.

external	list	OAuth Required. Conditional. This field is returned only if external values are set or included by using Opt In. The external field allows you to store app-specific metadata on tasks, including a gid that can be used to retrieve tasks and a data blob that can store app-specific character strings. Note that you will need to authenticate with Oauth to access or modify this data. Once an external gid is set, you can use the notation external:custom_gid to reference your object anywhere in the API where you may use the original object gid. See the page on Custom External Data for more details.

hearted	logical scalar	Deprecated - please use liked instead True if the task is hearted by the authorized user, false if not.

hearts	list	Deprecated - please use likes instead Array of likes for users who have hearted this task.

html_notes	character scalar	Opt In. The notes of the text with formatting as HTML.

is_rendered_as_separator	logical scalar	Opt In. In some contexts tasks can be rendered as a visual separator; for instance, subtasks can appear similar to sections without being true section objects. If a task object is rendered this way in any context it will have the property is_rendered_as_separator set to true.

liked	logical scalar	True if the task is liked by the authorized user, false if not.

likes	list	Array of likes for users who have liked this task.

memberships	list	Create-only. Array of projects this task is associated with and the section it is in. At task creation time, this array can be used to add the task to specific sections. After task creation, these associations can be modified using the addProject and removeProject endpoints. Note that over time, more types of memberships may be added to this property.

modified_at	datetime scalar	The time at which this task was last modified. Note: This does not currently reflect any changes in associations such as projects or comments that may have been added or removed from the task.

notes	character scalar	Free-form textual information associated with the task (i.e. its description).

num_hearts	integer scalar	Deprecated - please use likes instead The number of users who have hearted this task.

num_likes	integer scalar	The number of users who have liked this task.

num_subtasks	integer scalar	Opt In. The number of subtasks on this task.

start_at	date scalar	Date and time on which work begins for the task, or null if the task has no start time. This takes an ISO 8601 date string in UTC and should not be used together with start_on. Note: due_at must be present in the request when setting or unsetting the start_at parameter.

start_on	date scalar	The day on which work begins for the task , or null if the task has no start date. This takes a date with YYYY-MM-DD format and should not be used together with start_at. Note: due_on or due_at must be present in the request when setting or unsetting the start_on parameter.

assignee	list	A user object represents an account in Asana that can be given access to various workspaces, projects, and tasks.

assignee_section	list	The assignee section is a subdivision of a project that groups tasks together in the assignee's "My Tasks" list. It can either be a header above a list of tasks in a list view or a column in a board view of "My Tasks." The assignee_section property will be returned in the response only if the request was sent by the user who is the assignee of the task. Note that you can only write to assignee_section with the gid of an existing section visible in the user's "My Tasks" list.

custom_fields	list	Array of custom field values applied to the task. These represent the custom field values recorded on this project for a particular custom field. For example, these custom field values will contain an enum_value property for custom fields of type enum, a text_value property for custom fields of type text, and so on. Please note that the gid returned on each custom field value is identical to the gid of the custom field, which allows referencing the custom field metadata through the /custom_fields/custom_field-gid endpoint.

followers	list	Array of users following this task.

parent	list	The parent of this task, or null if this is not a subtask. This property cannot be modified using a PUT request but you can change it with the setParent endpoint. You can create subtasks by using the subtasks endpoint.

permalink_url	character scalar	A url that points directly to the object within Asana.

projects	list	Create-only. Array of projects this task is associated with. At task creation time, this array can be used to add the task to many projects at once. After task creation, these associations can be modified using the addProject and removeProject endpoints.

tags	list	Array of tags associated with this task. In order to change tags on an existing task use addTag and removeTag.

workspace	list	Create-only. The workspace this task is associated with. Once created, task cannot be moved to a different workspace. This attribute can only be specified at creation time.